+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Move `NullifierTree` and `BlockChain` from node to base (#1304).
 - Rename `ChainMmr` to `PartialBlockchain` (#1305).
 - Add safe `PartialBlockchain` constructor (#1308).
+- Generalized account-related inputs to the transaction kernel (#1311).
 
 ## 0.8.2 (2025-04-18) - `miden-proving-service` crate only